refactor: move memoization logic into own package
